// frontend/src/app/model/Groupe.ts
export class Groupe {

    group_code: Number;
    group_classif: Number;
    group_liblong: String;
    group_libcourt: String;
    group_siege: Number;
    group_adress1: String;
    group_adress2: String;
    group_email: String;
    group_web: String;
    group_teleph1: String;
    group_teleph2: String;
}